requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pydantic==2.9.0
python-multipart==0.0.9
pdfminer.six==20231228
